+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to HuduCommunityBot will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.13] - 2026-08-13
+
+### Changed
+
+* Lock community feed discussion threads after monitor posting
+
+## [1.5.12] - 2026-08-12
+
+### Changed
+
+* Fix duplicate Hudu release community link posts and lock release threads
+
 ## [1.5.11] - 2026-08-07
 
 ### Changed